BPTD-SoftwareAndDataSecurity/lab3/requirements.txt
python-docx==1.1.0
Pillow==10.2.0
